fix #2540 to make simple unclassified legend accordion be expandable
